requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
dataclasses-json==0.5.12
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
langchain==0.0.235
langsmith==0.0.7
marshmallow==3.19.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.11
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tqdm==4.65.0
typing-extensions==4.7.1
typing-inspect==0.9.0
urllib3==2.0.3
yarl==1.9.2
